FR47367: Location client API for External AP
(1) In location data struct, change size_t to uint32_t as size_t
is machine dependent and can not be sent as part of payload
to the other processor.
(2) Change LocIpc ifdef to match with the file name

CRs-fixed: 2367517

+++ location/LocationDataTypes.h
@@ -603,7 +603,8 @@ typedef struct {
 } GnssAidingData;
 
 typedef struct {
-    size_t size;             // set to sizeof(Location)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;           // set to sizeof(Location)
     LocationFlagsMask flags; // bitwise OR of LocationFlagsBits to mark which params are valid
     uint64_t timestamp;      // UTC timestamp for location fix, milliseconds since January 1, 1970
     double latitude;         // in degrees
@@ -619,7 +620,8 @@ typedef struct {
 } Location;
 
 typedef struct {
-    size_t size;          // set to sizeof(LocationOptions)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;          // set to sizeof(LocationOptions)
     uint32_t minInterval; // in milliseconds
     uint32_t minDistance; // in meters. if minDistance > 0, gnssSvCallback/gnssNmeaCallback/
                           // gnssMeasurementsCallback may not be called
@@ -627,32 +629,37 @@ typedef struct {
 } LocationOptions;
 
 typedef struct {
-    size_t size;
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;
     BatchingMode batchingMode;
 } BatchingOptions;
 
 typedef struct {
-    size_t size;
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;
     BatchingStatus batchingStatus;
 } BatchingStatusInfo;
 
 typedef struct {
-    size_t size;                            // set to sizeof(GeofenceOption)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                            // set to sizeof(GeofenceOption)
     GeofenceBreachTypeMask breachTypeMask;  // bitwise OR of GeofenceBreachTypeBits
     uint32_t responsiveness;                // in milliseconds
     uint32_t dwellTime;                     // in seconds
 } GeofenceOption;
 
 typedef struct {
-    size_t size;      // set to sizeof(GeofenceInfo)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;      // set to sizeof(GeofenceInfo)
     double latitude;  // in degrees
     double longitude; // in degrees
     double radius;    // in meters
 } GeofenceInfo;
 
 typedef struct {
-    size_t size;             // set to sizeof(GeofenceBreachNotification)
-    size_t count;            // number of ids in array
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;             // set to sizeof(GeofenceBreachNotification)
+    uint32_t count;            // number of ids in array
     uint32_t* ids;           // array of ids that have breached
     Location location;       // location associated with breach
     GeofenceBreachType type; // type of breach
@@ -660,7 +667,8 @@ typedef struct {
 } GeofenceBreachNotification;
 
 typedef struct {
-    size_t size;                       // set to sizeof(GeofenceBreachNotification)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                       // set to sizeof(GeofenceBreachNotification)
     GeofenceStatusAvailable available; // GEOFENCE_STATUS_AVAILABILE_NO/_YES
     LocationTechnologyType techType;   // GNSS
 } GeofenceStatusNotification;
@@ -794,7 +802,8 @@ typedef struct {
 } GnssSystemTime;
 
 typedef struct {
-    size_t size;                        // set to sizeof(GnssLocationInfo)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                        // set to sizeof(GnssLocationInfo)
     GnssLocationInfoFlagMask flags;     // bitwise OR of GnssLocationInfoBits for param validity
     float altitudeMeanSeaLevel;         // altitude wrt mean sea level
     float pdop;                         // position dilusion of precision
@@ -830,7 +839,8 @@ typedef struct {
 } GnssLocationInfoNotification;
 
 typedef struct {
-    size_t size;                           // set to sizeof(GnssNiNotification)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                           // set to sizeof(GnssNiNotification)
     GnssNiType type;                       // type of NI (Voice, SUPL, Control Plane)
     GnssNiOptionsMask options;             // bitwise OR of GnssNiOptionsBits
     uint32_t timeout;                      // time (seconds) to wait for user input
@@ -843,7 +853,8 @@ typedef struct {
 } GnssNiNotification;
 
 typedef struct {
-    size_t size;       // set to sizeof(GnssSv)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;       // set to sizeof(GnssSv)
     uint16_t svId;     // Unique Identifier
     GnssSvType type;   // type of SV (GPS, SBAS, GLONASS, QZSS, BEIDOU, GALILEO)
     float cN0Dbhz;     // signal strength
@@ -854,14 +865,16 @@ typedef struct {
 } GnssSv;
 
 typedef struct {
-    size_t size;             // set to sizeof(GnssConfigSetAssistanceServer)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;             // set to sizeof(GnssConfigSetAssistanceServer)
     GnssAssistanceType type; // SUPL or C2K
     const char* hostName;    // null terminated string
     uint32_t port;           // port of server
 } GnssConfigSetAssistanceServer;
 
 typedef struct {
-    size_t size;                               // set to sizeof(GnssMeasurementsData)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                               // set to sizeof(GnssMeasurementsData)
     GnssMeasurementsDataFlagsMask flags;       // bitwise OR of GnssMeasurementsDataFlagsBits
     int16_t svId;
     GnssSvType svType;
@@ -885,7 +898,8 @@ typedef struct {
 } GnssMeasurementsData;
 
 typedef struct {
-    size_t size;                          // set to sizeof(GnssMeasurementsClock)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                          // set to sizeof(GnssMeasurementsClock)
     GnssMeasurementsClockFlagsMask flags; // bitwise OR of GnssMeasurementsClockFlagsBits
     int16_t leapSecond;
     int64_t timeNs;
@@ -899,35 +913,41 @@ typedef struct {
 } GnssMeasurementsClock;
 
 typedef struct {
-    size_t size;                 // set to sizeof(GnssSvNotification)
-    size_t count;                // number of SVs in the GnssSv array
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                 // set to sizeof(GnssSvNotification)
+    uint32_t count;                // number of SVs in the GnssSv array
     bool gnssSignalTypeMaskValid;
     GnssSv gnssSvs[GNSS_SV_MAX]; // information on a number of SVs
 } GnssSvNotification;
 
 typedef struct {
-    size_t size;         // set to sizeof(GnssNmeaNotification)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;         // set to sizeof(GnssNmeaNotification)
     uint64_t timestamp;  // timestamp
     const char* nmea;    // nmea text
-    size_t length;       // length of the nmea text
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t length;       // length of the nmea text
 } GnssNmeaNotification;
 
 typedef struct {
-    size_t size;                 // set to sizeof(GnssDataNotification)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                 // set to sizeof(GnssDataNotification)
     GnssDataMask  gnssDataMask[GNSS_LOC_MAX_NUMBER_OF_SIGNAL_TYPES];  // bitwise OR of GnssDataBits
     double        jammerInd[GNSS_LOC_MAX_NUMBER_OF_SIGNAL_TYPES];     // Jammer Indication
     double        agc[GNSS_LOC_MAX_NUMBER_OF_SIGNAL_TYPES];           // Automatic gain control
 } GnssDataNotification;
 
 typedef struct {
-    size_t size;         // set to sizeof(GnssMeasurementsNotification)
-    size_t count;        // number of items in GnssMeasurements array
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;         // set to sizeof(GnssMeasurementsNotification)
+    uint32_t count;        // number of items in GnssMeasurements array
     GnssMeasurementsData measurements[GNSS_MEASUREMENTS_MAX];
     GnssMeasurementsClock clock; // clock
 } GnssMeasurementsNotification;
 
 typedef struct {
-    size_t size;  // set to sizeof(GnssConfig)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;  // set to sizeof(GnssConfig)
     GnssConfigFlagsMask flags; // bitwise OR of GnssConfigFlagsBits to mark which params are valid
     GnssConfigGpsLock gpsLock;
     GnssConfigSuplVersion suplVersion;
@@ -942,7 +962,8 @@ typedef struct {
 } GnssConfig;
 
 typedef struct {
-    size_t size;                        // set to sizeof
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                        // set to sizeof
     bool                                mValid;
     Location                            mLocation;
     double                              verticalAccuracyMeters;
@@ -952,7 +973,8 @@ typedef struct {
 } GnssDebugLocation;
 
 typedef struct {
-    size_t size;                        // set to sizeof
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                        // set to sizeof
     bool                                mValid;
     int64_t                             timeEstimate;
     float                               timeUncertaintyNs;
@@ -960,7 +982,8 @@ typedef struct {
 } GnssDebugTime;
 
 typedef struct {
-    size_t size;                        // set to sizeof
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                        // set to sizeof
     uint32_t                            svid;
     GnssSvType                          constellation;
     GnssEphemerisType                   mEphemerisType;
@@ -972,7 +995,8 @@ typedef struct {
 } GnssDebugSatelliteInfo;
 
 typedef struct {
-    size_t size;                        // set to sizeof
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size;                        // set to sizeof
     GnssDebugLocation                   mLocation;
     GnssDebugTime                       mTime;
     std::vector<GnssDebugSatelliteInfo> mSatelliteInfo;
@@ -1047,7 +1071,8 @@ typedef std::function<void(
    collectiveResponseCallback is called for every geofence API call.
    ids array and LocationError array are only valid until collectiveResponseCallback returns. */
 typedef std::function<void(
-    size_t count, // number of locations in arrays
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t count, // number of locations in arrays
     LocationError* errs, // array of LocationError associated to the request
     uint32_t* ids // array of ids to be associated to the request
 )> collectiveResponseCallback;
@@ -1063,7 +1088,8 @@ typedef std::function<void(
    batchingCallback is called when delivering locations in a batching session.
    broadcasted to all clients, no matter if a session has started by client */
 typedef std::function<void(
-    size_t count,      // number of locations in array
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t count,      // number of locations in array
     Location* location, // array of locations
     BatchingOptions batchingOptions // Batching options
 )> batchingCallback;
@@ -1138,7 +1164,8 @@ typedef std::function<void(
 )> locationApiDestroyCompleteCallback;
 
 typedef struct {
-    size_t size; // set to sizeof(LocationCallbacks)
+    // do not use size_t as data type for size_t is architecture dependent
+    uint32_t size; // set to sizeof(LocationCallbacks)
     capabilitiesCallback capabilitiesCb;             // mandatory
     responseCallback responseCb;                     // mandatory
     collectiveResponseCallback collectiveResponseCb; // mandatory
